requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
pandas>=2.2.0
pandas-ta>=0.3.14b
pyyaml>=6.0.1
python-telegram-bot>=21.0
apscheduler>=3.10.4
